libsel4platsupport: fix seL4_DebugPutChar detection
Serial console support can fall back to using seL4_DebugPutChar. But
it does this based on the kernel's CONFIG_DEBUG_BUILD config parameter
which is wrong--the support is actually based on CONFIG_PRINTING which
is implictly enabled when CONFIG_DEBUG_BUILD is set. This is confirmed
by the fact the actual usage of seL4_DebugPutChar is conditional on
CONFIG_PRINTING. Replace CONFIG_DEBUG_BUILD use with CONFIG_PRINTING.

This has the added benefit of fixing sel4test serial server test cases
when a release build is used.

Bug: 250071553

Change-Id: I50cc2c898be43ff95f4c8fa53c6ed7bb3446bca2